add App contains router
include App.test.js
