+++ frontend/src/index.js
@@ -1,70 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client'
-import {
-  createBrowserRouter,
-  RouterProvider,
-} from "react-router-dom";
-
-import Root from "./routes/root";
-import Index from './routes';
-import ErrorPage from './error-page';
-
-import { action as authAction } from './routes/auth';
-import { action as logoutAction } from './routes/logout';
-
-import Bikes, {
-  loader as bikesLoader,
-} from './routes/bikes';
-
-import Riders, {
-  loader as ridersLoader,
-} from './routes/riders';
 
 import reportWebVitals from './reportWebVitals';
 import 'bootstrap/dist/css/bootstrap.css';
-import Authentication from './routes/auth';
-import { tokenLoader } from './utils/auth';
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root />,
-    loader: tokenLoader,
-    errorElement: <ErrorPage />,
-    children: [
-      {
-        errorElement: <ErrorPage />,
-        children: [
-          { index: true, element: <Index /> },
-          {
-            path: 'auth',
-            element: <Authentication />,
-            action: authAction,
-          },
-          {
-            path: 'logout',
-            action: logoutAction,
-          },
-          {
-            path: "bikes",
-            element: <Bikes />,
-            loader: bikesLoader,
-          },
-          {
-            path: "riders",
-            element: <Riders />,
-            loader: ridersLoader,
-          },
-        ],
-      },
-    ],
-  },
-]);
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
   <React.StrictMode>
-    <RouterProvider router={router} />
+    <App />
   </React.StrictMode>
 );
